+++ CHANGELOG.md
@@ -3,10 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [🔖 Unreleased] - 17/06/2021
+## [🔖 Unreleased]
 ### ✨ Added
 - Adds casting of element types.
-- `Contains`, `Distinct` and `Reverse` methods.
+- `Contains`, `Distinct`, `Single` and `Reverse` methods.
+
+### 🙌 Improvements
+- `CLinqCollection<TElement>` can now be constructed with initializer list.
 
 ## [🔖 [0.1.0]](https://github.com/MattBolitho/CLinq/releases/tag/CLinq-0.1) - 17/06/2021
 ### ✨ Added